fix: pretty print new `package.json` files
When a `package.json` is read via pkg-types, we store the indent level
for later use and apply it when writing the same file.

However, if we store an object which wasn't originally read via
pkg-types, we don't have this information and will write the file with
no indentation (i.e. the output of `JSON.stringify(obj)`).

This fix changes the behaviour such that read/writes still retain the
indentation information, but new files default to `indent: 2`.